requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9
razorpay>=1.3.0
python-dotenv>=1.0.0
Pillow>=10.0.0
gunicorn>=21.2.0